update gitignore and tsconfig.json + further refactoring
