--- index.ts
@@ -1,2 +1,0 @@
-export * from "./src/extension";
-export * from "./src/contracts";
