+++ package.json
@@ -17,7 +17,9 @@
   "keywords": [
     "tiptap",
     "annotation",
-    "comments"
+    "comments",
+    "wysiwyg",
+    "nlp"
   ],
   "dependencies": {
     "@tiptap/core": "^2.1.7",
