+++ src/extension/pm/annotation-decoration.ts
@@ -1,5 +1,5 @@
 import { Decoration } from "@tiptap/pm/view";
-import { Annotation } from "../../contracts/annotation";
+import { Annotation } from "../../contracts";
 
 export class AnnotationDecoration<K> implements Annotation<K> {
   private decoration!: any;
